Introduce new 'container_mounts' config key
The older 'container_volumes' key is now deprecated. It was backwards, in the
sense that it was a {host_dir: container_dir} mapping, whereas container_dirs
are unique, and host_dirs are not, which means that container_dir should be
the dictionary keys, and host_dirs should be the values. This is the case with
the new 'container_mounts' configuration key.

Fixes: https://github.com/brainvisa/casa-distro/issues/91