don't install pixi env at setup
because pixi envs are not relocatable and at setup the install dir is /casa/setup/install instead of /casa/host/install.
Thus we just create the pixi.toml file and some hooks, which are executed at the next casa_distro start.
The cache directory is also shared witht the host user, if on the same filesystem.